added data-ignore-validation attribute check
added data-ignore-validation attribute. if the attribute is passed with the value 'true', it will bypass the validation for the phone number format, as there could be a case where one would not want to validate the format of phone number and would simply let user enter number in what ever way they want.